example - initialize rugged-donut
note: Link to original python code in code

The next few commits are going to be about designing the end goal

With this next attempt at tackling the problem, would be nice to flesh
out what exactly the mapping between clojure syntax to python will be.

Previous attempt was a slog because almost no planning was done,
implementation and design were happening at the same time which got
really messy when for example I hadn't considered well enough how to
handle all edge cases for data types other than primitives.  The
`clj->py-value` logic swapping between raw strings and clojure values
was kind of a bit of everywhere depending on the particular parse-method
which got really gross really quickly.

I searched for a simple example of a (relatively) typical script and
this seems like a good place to start, it seems to consists of certain
operations:

1. creating objects
2. applying created textures/modifiers to objects
3. keyframe animation?

There are alot of things to consider