example - global vars + calling custom functions
This covers lines 6-13. the initialization code of the script, it's a
pretty 1-1 mapping from the clojure code to python code

The import on `utils` was moved to a require as it seems more likely
thatcustom scripts will also be written with donkeyblend (most likely?),
aka a clojure script, thus using `require` instead. Though using
`import` still works (for custom scripts written in Python).