example - operations that rely on mutations
This diff covers lines 29-35.

This took much more deliberation

    for frame in range(1, num_frames):
        ...
        empty.location = (x, y, z)
        empty.keyframe_insert(data_path="location", index=-1, frame=frame)

The way this method `keyframe_insert` uses `empty.location` inside the
function definition and the only way of having a construct like this is
to have mutation.

Clojure doesn't really have idiomatic langauge construct that can handle
a usecase like this. The best I can think of some sort of combination of
records + an atom...which is pretty complicated in my opinion.

Another option is to do something like this

  (obj (.location [x y z]))
  (obj (.keyframe_insert {:data_path "locations"
                          :index -1
                          :frame frame})))

Where the `.location` method is just another property. This looks pretty
clean but the downside is that there's alot more responsibility on the
language to be able to interpret what `.location` and `.keyframe_insert`
are. Mixing attributes and methods will require some sort of extra logic
to differentiate the two back out at runtime. It's possible as we can do
some sort of `try...catch` logic (super expensive) or to wrap everything
in some sort of `if type(...)` statement which dispatches by type.

The approach that was adopted ended up being using type annotations as
it is very much like the approach above except the user/client provides
the extra information that this is a class attribute. Also, it looks
cleaner and idiomatic Clojure uses type hints similarly.