reader returns all objects in string
NOTE: This is more of a hacky approach to fixing this

The way that `edn/read-string` works it returns the first object in a
string.

You can solve this problem a bit more elegantly by using the `read`
function which uses Java's `PushbackReader` to look forward but requires
a bit more boilerplate.

For now however, this suits our requirements.