reader option initialization
Skeleton code for the reader options. A little bit about the opts:

- `:readers` is passing in functions to read custom symbols
- `:default` is the default case for if none of the readers match. Not
too sure how this works yet, the documentation in `cljs.reader/read`
seems to indicate that it's a function of two arguments `tag` and
`value`...not too sure what `value` is supposed to be yet. Something to
keep in mind
-  `:eof` is the value that gets returned for an EOF character. `nil` is
already set as default but looking at the implementation, it needs to be
explicitly passed if a map is being passed in as the first argument.

Tests still pass